WIP - make common hashing utilities reusable
